Extended drift correction so that when you plot the results of calibration in Experteyes, the secondary calibration uses the LAST drift correction value (if there is one) to correct the secondary calibration.
The idea is that this can be used when you are going through the trials to see if you caught the last drift. The assumption is that drifts occur as discete events that can be tracked between trials by looking at a fixation spot and setting a drift set to that fixation spot.

Of course you could set a drift correction at the start of the secondary calibration sequence at the end of the experiment but that would be kind of silly and only tell you if you got drift during the secondary calibration.