Fixed a problem with the drift correction so that all drift correction uses the sync frame number, not the eye frame number. This makes it generic regardless of whether the eye sync frame preceeds or follows the scene sync frame.
Tested both the movie export and the data export.

It is important for the data export that the secondary eye calibration follows the experiment (i.e. is at the end). Any drift correction that is applied during the experiment also affects the secondary calibration. However, if the secondary calibraiton is set as the one to use in matlab (this is rare) the use of drift correction doesn't make sense.

This is a stable release, ExpertEyes 2.2.